// frontend/src/core/api/generated/model/taskView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { TaskPerson } from './taskPerson'
import type { TaskViewPriority } from './taskViewPriority'
import type { TaskViewStatus } from './taskViewStatus'

export interface TaskView {
  id: string
  siteId: string
  title: string
  /** @nullable */
  note?: string | null
  status: TaskViewStatus
  priority: TaskViewPriority
  /** @nullable */
  dueDate?: string | null
  assignee?: TaskPerson
  createdBy: TaskPerson
  /** @nullable */
  postId?: string | null
  createdAt: string
  /** @nullable */
  completedAt?: string | null
}